fix: address Copilot review comments
- Move hazeEffect overload to dev.chrisbanes.haze package for v1 import compatibility
- Fix ReplaceWith to preserve receiver (remove hard-coded Modifier)
- Remove unused imports from MigrationAliases.kt
- Clean up HazeEffectScope.kt imports
